backend/requirements.txt
Django==5.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-filter==23.5
Pillow==10.4.0
PyMySQL==1.1.0
python-dotenv==1.0.0
openpyxl==3.1.2
